FIX: float64 promotion in kernels causes NaN forces on GPUs without float64 XLA
laplace_kernel used Python float literal 1.0 (float64) to divide a float32
numpy array, promoting to float64. gradient_kernel used Python 1j (complex128)
similarly. When jax_cosmo enables x64 at import, this causes complex128 FFTs
on GPUs without full float64 XLA support → silent NaN in all force computations.

Fix: use np.float32(1.0) and np.complex64(1j) in kernels.py so everything
stays float32/complex64 regardless of x64 config.
Add JAX_ENABLE_X64=0 env guard to training scripts as belt-and-suspenders.

+++ jaxpm/kernels.py
@@ -34,15 +34,15 @@ def gradient_kernel(kvec, direction, order=1):
       Complex kernel
     """
     if order == 0:
-        wts = 1j * kvec[direction]
+        wts = np.complex64(1j) * kvec[direction]   # stay complex64
         wts = jnp.squeeze(wts)
         wts[len(wts) // 2] = 0
         wts = wts.reshape(kvec[direction].shape)
         return wts
     else:
         w = kvec[direction]
-        a = 1 / 6.0 * (8 * jnp.sin(w) - jnp.sin(2 * w))
-        wts = a * 1j
+        a = np.float32(1 / 6.0) * (8 * jnp.sin(w) - jnp.sin(2 * w))
+        wts = a * np.complex64(1j)   # stay complex64
         return wts
 
 
@@ -61,7 +61,7 @@ def laplace_kernel(kvec):
     kk = sum(ki**2 for ki in kvec)
     mask = (kk == 0).nonzero()
     kk[mask] = 1
-    wts = 1.0 / kk
+    wts = np.float32(1.0) / kk   # float32 literal avoids float64 promotion
     imask = (~(kk == 0)).astype(int)
     wts *= imask
     return wts
